Initial commit of unreson
It works but there is some special use requirements for adding new
properties to a contained object:

Given "obj.state.a.b = {}", you cannot provide "obj.state.a.b.c = 1",
but rather you must provide "obj.state.a.b = { c: 1 }". Once this object
has been established, it _should_ work. Unsure if there is a way to get
around this or if I made an error in my implementation of Proxies...