Only return a state proxy via get()
This is a change from the previous behaviour, where we would set the
internal _state to a Proxy when set() was used. Now we only return a
Proxy when get() is called. This allows for 3rd-party libraries to
access the _state property directly for analysis without incurring the
costs of using proxies for very large objects.

A getter for the internal state should likely be implemented so such
library usage does not necessitate using the internal state directly.